Restructure query: Change calculation of referred trading volume  (#25)
REPLACES #23 

Previously, the query calculated the referrer trading volume by adding all trading volume that was traded with a specific appData(and hence referral). But this has the weakness that everyone will just switch the appData and then alter on referrer to a friend.

A more effective mechanism is to observe the first trade from a user, inspect the appData used, derive the referrer from it and then account later on all trading volume generated by this user to the referral, no matter whether a new appData is used. This will disallow users to change the referral later to their friend.

Testplan:
Visit the two queries affected and observe that they are still working:

https://dune.xyz/queries/135804
https://dune.xyz/queries/157348