Get newest app data from solveable orders (#32)
Currently, we would get the app-data from the dune-queries. Though going that route takes additional time. It's much easier to always get the newest dune data from the solvable orders, instead of waiting for the dune queries. This PR implements this fetching method, additional to the existing one.

Later on, I would like to make a new PR to the gp-v2-services, that just provides all app_data and then we don't need the dune query at all for this data.

Testplan:
unit tests